move exceptions to utils
