+++ reconcile/utils/jump_host.py
@@ -11,7 +11,7 @@ from sshtunnel import SSHTunnelForwarder
 from reconcile.utils import gql
 from reconcile.utils.secret_reader import SecretReader
 
-from reconcile.exceptions import FetchResourceError
+from reconcile.utils.exceptions import FetchResourceError
 
 # https://www.iana.org/assignments/service-names-port-numbers/service-names-port-numbers.xhtml
 DYNAMIC_PORT_MIN = 49152
