+++ reconcile/utils/terrascript_client.py
@@ -82,7 +82,7 @@ from reconcile.utils.git import is_file_in_git_repo
 from reconcile.github_org import get_config
 from reconcile.utils.oc import StatusCodeError
 from reconcile.utils.gpg import gpg_key_valid
-from reconcile.exceptions import FetchResourceError
+from reconcile.utils.exceptions import FetchResourceError
 from reconcile.utils.elasticsearch_exceptions \
     import (ElasticSearchResourceNameInvalidError,
             ElasticSearchResourceMissingSubnetIdError,
